+++ src/basetext.c
@@ -1,4 +1,4 @@
-#include "stb_text.h"
+#include "utils.h"
 #include <stdlib.h>         // Required for: malloc(), free()
 #include <stdio.h>          // Required for: vsprintf()
 #include <string.h>         // Required for: strcmp(), strstr(), strcpy(), strncpy() [Used in TextReplace()], sscanf() [Used in LoadBMFont()]