+++ changelog.md
@@ -1,3 +1,4 @@
-# V1.1.2
-- 重新添加`rcamera`文件。
-- 解决`WaitTime`函数延时精度问题。
+# V1.1.3
+- 因为要使用`TextFormat`函数而将`stb_text`重新并入`raylib.h`。
+- 删除`DrawFPS`函数，`GetFPS`函数中取消滑动平均处理。
+- 添加`GetCPUusage`函数。